+++ ledger/ledger-api-tests/suites/src/main/scala/com/daml/ledger/api/testtool/suites/v1_8/ValueLimitsIT.scala
@@ -6,8 +6,7 @@ package com.daml.ledger.api.testtool.suites.v1_8
 import com.daml.ledger.api.testtool.infrastructure.Allocation._
 import com.daml.ledger.api.testtool.infrastructure.Assertions._
 import com.daml.ledger.api.testtool.infrastructure.LedgerTestSuite
-import com.daml.ledger.client.binding.Primitive.Party
-import com.daml.ledger.test.model.Test.{DummyWithAnnotation, WithList, WithMap}
+import com.daml.ledger.test.model.Test.DummyWithAnnotation
 
 import scala.concurrent.Future
 import scala.util.Random
@@ -41,62 +40,4 @@ final class ValueLimitsIT extends LedgerTestSuite {
       ()
     }
   })
-
-  test(
-    "VLLargeMapInContract",
-    "Create a contract with a field containing large map",
-    allocate(SingleParty),
-    8.0,
-  )(implicit ec => { case Participants(Participant(ledger, alice)) =>
-    val elements = (1 to 10000).map(e => (f"element_$e%08d", alice)).toMap
-    for {
-      contract <- ledger.create(alice, WithMap(alice, elements))
-      _ <- ledger.exercise(alice, contract.exerciseWithMap_Noop)
-    } yield {
-      ()
-    }
-  })
-
-  test(
-    "VLLargeMapInChoice",
-    "Exercise a choice with a large map",
-    allocate(SingleParty),
-  )(implicit ec => { case Participants(Participant(ledger, alice)) =>
-    val elements = (1 to 10000).map(e => (f"element_$e%08d", alice)).toMap
-    for {
-      contract <- ledger.create(alice, WithMap(alice, Map.empty[String, Party]))
-      _ <- ledger.exercise(alice, contract.exerciseWithMap_Expand(_, elements))
-    } yield {
-      ()
-    }
-  })
-
-  test(
-    "VLLargeListInContract",
-    "Create a contract with a field containing large list",
-    allocate(SingleParty),
-  )(implicit ec => { case Participants(Participant(ledger, alice)) =>
-    val elements = (1 to 10000).map(e => f"element_$e%08d")
-    for {
-      contract <- ledger.create(alice, WithList(alice, elements))
-      _ <- ledger.exercise(alice, contract.exerciseWithList_Noop)
-    } yield {
-      ()
-    }
-  })
-
-  test(
-    "VLLargeListInChoice",
-    "Exercise a choice with a large list",
-    allocate(SingleParty),
-  )(implicit ec => { case Participants(Participant(ledger, alice)) =>
-    val elements = (1 to 10000).map(e => f"element_$e%08d")
-    for {
-      contract <- ledger.create(alice, WithList(alice, List.empty[String]))
-      _ <- ledger.exercise(alice, contract.exerciseWithList_Expand(_, elements))
-    } yield {
-      ()
-    }
-  })
-
 }
