+++ ledger/test-common/src/main/daml/model/Test.daml
@@ -5,7 +5,6 @@
 
 module Test where
 
-import DA.Map qualified as M
 import DA.Time
 
 import Iou()
@@ -732,40 +731,3 @@ template WithKeyDivulgenceHelper
       controller withKeyOwner
       do _ <- fetch withKeyToFetch
          return ()
-
-template WithMap
-  with
-    owner : Party
-    items : M.Map Text Party
-  where
-    signatory owner
-    nonconsuming choice WithMap_Noop : ()
-      controller owner
-      do pure ()
-
-    choice WithMap_Expand : ContractId WithMap
-      with
-        newItems : M.Map Text Party
-      controller owner
-      do
-        create this with
-          items = M.union items newItems
-
-
-template WithList
-  with
-    owner : Party
-    items : [Text]
-  where
-    signatory owner
-    nonconsuming choice WithList_Noop : ()
-      controller owner
-      do pure ()
-
-    choice WithList_Expand : ContractId WithList
-      with
-        newItems : [Text]
-      controller owner
-      do
-        create this with
-          items = items ++ newItems
