+++ ledger/test-common/src/main/daml/semantic/Limits.daml
@@ -1,0 +1,43 @@
+-- Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+-- SPDX-License-Identifier: Apache-2.0
+
+module Limits where
+
+import DA.Map qualified as M
+
+template WithMap
+  with
+    owner : Party
+    items : M.Map Text Party
+  where
+    signatory owner
+    nonconsuming choice WithMap_Noop : ()
+      controller owner
+      do pure ()
+
+    choice WithMap_Expand : ContractId WithMap
+      with
+        newItems : M.Map Text Party
+      controller owner
+      do
+        create this with
+          items = M.union items newItems
+
+
+template WithList
+  with
+    owner : Party
+    items : [Text]
+  where
+    signatory owner
+    nonconsuming choice WithList_Noop : ()
+      controller owner
+      do pure ()
+
+    choice WithList_Expand : ContractId WithList
+      with
+        newItems : [Text]
+      controller owner
+      do
+        create this with
+          items = items ++ newItems
